Consolidated all error codes into constants.
Errors are so inconsistently used across the OAuth API that it
doesn't make a lot of sense to separate them by type. This patch
moves all throwable exceptions (and their codes) into a single
class, and adds an easily queryable errorcode utility class
for all the strings that should be represented.