Consolidated AnalyzerDef
Turns out that @AnalyzerDef only declares an analyzer, and doesn't
apply it. As such, it basically means that this didn't actually do
anything.

This patch moves the analyzer defenition into AbstractEntity and
removes it from the other entities that it was using. We'll have
to apply them judiciously as we start building up our search indexes.