OAuth2 Post-Matching filter
This patch adds the @OAuth annotation, which adds token
validation logic to the selected resource path, as well as
the postmatching filter which enables this logic. The user
resource has been updated to require a token with the "user"
scope.

This implementation still depends on access to the database.
Eventually, it should be replaced with a caching implementation
that validates the token independently.