Add kangaroo namespace to scope. (#133)
When creating a new application on a kangaroo server, we would
like to be able to grant the users of that application limited
access to the kangaroo API. In order to accomplish this, we will
make the kangaroo:scopes available directly to each created application;
and to avoid namespace collisions, prefix it with "kangaroo:". Note
that any user, not a member of the admin application, who attempts to
access data on the kangaroo API, should only be able to read data
related to themselves, and should only be able to modify their own
identities.