Converted DatabaseManager into ExternalResource
In order to separate concerns a little, I've converted the DatabaseManager,
which was explicitly managed, into a JUnit4 @ClassRule based resource.
This encapsulates it in a slightly more recognizable way. Furthermore,
rather than trying to generate our own SessionFactory from whatever
configuration we can glean, we're instead registering the Test Instance
with the injection context, so that we'll receive an instance of the
actual Application Contexts' SessionFactory. This should mitigate
race conditions while accessing the search indexes.

This patch identifies a bug caused by improper search index entity
linking in users/identies. It has been fixed.