Permitted token scopes are now restricted by role.
In the flows where a specific user is making a token request
(all flows except for client_credentials), the list of scopes
a request may issue is now restricted by the scopes permitted
to the role assigned by the user. This patch includes
modifications to the Validationutil, associated tests, as
well as the TestAuthenticator.

Closes #105