Collapsed project structure (#147)
Since it's possible to publish- and depend on - test-jar's, there's
no real point in maintaining so many different subprojects. This
patch collapses kangaroo-test, kangaroo-database, and
kangaroo-database-test, into kangaroo-common, and removes several
classes deemed to be superfluous as a result. It also removes
the debug flag from the logback configuration, in order to reduce
the size of our logs.