Ported knet utilities into kangaroo-common
It's a bit easier to manage this code inside of the lifecycle of this
application, rather than having to rerelease the separate version
whenever a new use case arises or the old ones change. In this case,
it was prompted by the load order of the search index listener vs.
the FirstRunContainer; As the search indexes cannot be built without
an existing schema.