Created JUnit rules for hibernate session creation and test data loading. (#198)
These two rules allow us to control whether our database sessions, and our
test data, are loaded either at a class or at a test level. There's a
tradeoff: In order to be useful and accessible during the test, they
must open their own connections to the database; using these in combination
with the DatabaseResource means that each test requires the creation of
4 connections to the database. Thankfully, these ones are able to dispose
of themselves cleanly in the finally{} blocks.

A side effect of this change is that the hibernate configuration can also
be configured using system properties. This should come in handy when
we finally wrap it all into a docker container.