Migraded oauth2 api to use Database Rules (#204)
This patch updates the OAuth2 project to use the Database Rules
created in a previous patch. It should improve the speed of these
tests as the database is only constructed once.

Note that to get these to pass, we have to make sure that the
hibernate rule reaches out to the database to grab a connection,
before we test for lingering DB sessions. As hibernate connections
are only constructed on an as-needed basis, not doing this
causes the tests to fail.