Fix two bugs in the ApplicationBuilder
This patch fixes two bugs. First, we copy any context that is returned
from the application builder, so that the ApplicationBuilder may be
reused without poisoning an already-built context.

The second bug uses the user identity passed into the method, rather
than the one from the context, to create a new token.