User Identity is now persisted with reverse reference.
Something in the Hibernate Search SDK has issues when an
entity, with child search index paths, that does not have its
parent association set, is persisted. In this case, the
parent's search index is not updated with the child's child
indexes. This is fixed by explicitly creating the parent-child
link from both the parent's and the child's perspective.