Entity path creation in test harness is now done by entity
With this patch, we introduce a new method that builds URI paths
for the test harness. In subresources, it is often necessary to
build a path related to the parent entity, which cannot be successfully
inferred simply from an ID. Thus, we've introduced getUrlByEntity, which
allows this.

We've also included a battery of checkstyle fixes.