Switched from WAR to JAR packaging (#260)
This allows us to launch the server using java -jar instead of
running it inside a tomcat servlet container. It includes a ServerFactory
which wraps common configuration options and setup for a grizzly-based,
multi ResourceConfig servlet host. Classes with main() methods have also
been added, as well as an assembly descriptor that collates all services.