Switched datetime columns to bigint
Different databases do not implement date manipulation functions in the same
way, or with the same precision, so we have to rely on a raw integer type instead.
This patch converts the underlying data type to bigint (essentially a timestamp),
and adds a type converter to manage the conversion into a Calendar type.