HTTPS by default (#279)
* Created testing p12 certificate

A simple certificate with full authentication, to test that we can
start an https server.

* Added configuration options for the keystore

These options permit setting a keystore path, password, certificate
alias and certificate password.

* Added SSL Certificate Handling

This patch forces the server into a 'secure-only' mode, where
an SSL certificate is required for operation. In cases where no certificate
is provided, the server will generate its own. Commandline parameters have
been provided to enable providing a keystore to the runtime, from which
an aliased key will be extracted for proper SSL operation.