Removed redirect from KangarooException (#287)
The common exception does not have a redirect use case- that is isolated
to the OAuth2 service, and should be handled there. This patch removes
all references to Redirect from the base KangarooException, and adds
a new RedirectingException type int he OAuth2 application which adds the
necessary transforms to provide a compliant error code, targeted at a
valid redirect return.