Created HtmlApplicationHttpHandler for root server.
By providing the --kangaroo.html_app_root commandline parameter,
a user may instruct the server to host an HTML application at the root
of the server path. We define this  as a rich javascript client, able
to manage its own routing via the HTML5 history API.

In order to ensure consistent usage of security headers, we also extracted
these values into a common configuration list, which is used both in the
newly created handler, and in the SecurityFeature.

Closes #289