Implicit flow now manages refresh tokens via the browser session.
In order to provide a seamless authn flow for long-lasting implicit
flow sessions (while still permitting short-lived tokens), the
/authorize endpoint now maintains a session cookie used as a
key into a list of refresh tokens. If a user asks for a new token
using the implicit flow, and a valid refresh token is attached to
the browser session (and they both are valid) for the requested client,
the resulting redirect will return immediately to the provided redirect
endpoint with a new token.

If a refresh token is _not_ found, the 302 will still point to the
endpoint of the IdP.