Created HttpSession entity
HTTP sessions are used by the Implicit type handler in lieu of a refresh
token. Here, we create a HibernateEntity which matches the Grizzly Session
instance specification, so we can maintain state via the database. This
session's function is to act solely as a reference for refresh tokens,
and it makes use of the SecureRandomIdGenerator to ensure sufficient
randomness.