ValidationUtil now only grants those scopes that exit
We're switching the behavior here, causing our ValidationUtil to
return the list of valid scopes that may be issued to a user, and filtering
out those that may not. This allows us to create 'blanket' authz requests
for all available scopes, and get back the subset that has actually been granted
to the user.