Corrected Jacoco coverage. (#338)
In a previous patch the jacoco configuration started to silently fail, due to a
missed commandline instruction. This patch adds it back in, and adds missing tests.
It also includes an update to mockito2, which can now mock final classes. This
makes Powermock obsolete, and we'll be removing it in subsequent patches.