Changing the session id updates stored timestamps. (#344)
Grizzly's underlying session maintenance - when provided only with a
maxage - will use Systen,currentTimeMillis() to set the MaxAge and ExpireDate
times in the cookie. This creates a mismatch with the database implementation,
which maintained the current and modified dates from the previous session,
and thus would cause the session to age out before the cookie that was provided
to the browser.

This patch removes the query that updates the created and modified dates in
the newly created http session record, bringing the cookie age on the server
in line with the cookie age on the client. It also changes the version of the
cookie to "1", to ensure that grizzly will send the MaxAge date, rather than the
Expires on.