Working Directory (#378)
This pull request creates the notion of a "Working Directory" nominally at /var/run/kangaroo that contains runtime resources for the server. These resources (at this time) only include the generated h2 database (if running in h2 mode) and the generated self-signed server certificate store (if none was provided), however it can be expanded in the future for other purposes (lucene indexes for instance).

The most notable improvement is that state will be maintained between restarts with no special configuration. The h2 database will remain, as will the certificate.