De-generify the token service injection pattern. (#395)
Previously, we merely accepted all parameters into the endpoint, accepting that the downstream handler
would know what to do with it. This is not really easy to handle for swagger however, as it doesn't
provide an effective way to annotate a method with parameters that aren't explicitly declared.
Rather than try to write our own annotations, we've simply explicitlye declared all possible input
parameters into the entity, followed by making the injection points explicit, rather than relying
on ITokenRequestHandler.