Change default working directory
This is a breaking change for the default configuration, though perhaps a welcome one.
The default working_dir of kangaroo was, by default, set to a directory that is not
automatically accessible to an average user. While our desired directory in a server
deployment should be /var/lib/kangaroo, for the sake of ease-of-use for new users,
we're switching this to `pwd` instead of one that would require sudo.

Closes #431